+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "acvm-backend-barretenberg"
 description = "An ACVM backend which allows proving/verifying ACIR circuits against Aztec Lab's Barretenberg library."
-version = "0.10.0"
+version = "0.11.0"
 authors = ["The Noir Team <team@noir-lang.org>"]
 edition = "2021"
 rust-version = "1.66"
@@ -13,7 +13,7 @@ license = "MIT OR Apache-2.0"
 crate-type = ["cdylib", "lib"]
 
 [dependencies]
-acvm = { version = "0.21.0", features = ["bn254"] }
+acvm = { version = "0.22.0", features = ["bn254"] }
 bincode = "1.3.3"
 bytesize = "1.2"
 reqwest = { version = "0.11.16", default-features = false, features = [
