--- src/acvm_interop/mod.rs
@@ -1,8 +1,0 @@
-use crate::Barretenberg;
-
-mod common_reference_string;
-mod proof_system;
-mod pwg;
-mod smart_contract;
-
-impl acvm::Backend for Barretenberg {}
