+++ src/barretenberg/schnorr.rs
