+++ src/common_reference_string.rs
@@ -1,6 +1,6 @@
 use acvm::{acir::circuit::Circuit, async_trait, CommonReferenceString};
 
-use crate::{composer::Composer, BackendError, Barretenberg};
+use crate::{barretenberg::composer::Composer, BackendError, Barretenberg};
 
 // TODO(#185): Ensure CRS download works in JS
 #[async_trait(?Send)]
